initial refactoring for issues #25 and #27
