+++ Migration1x-2x.md
@@ -1,0 +1,11 @@
+# Migration
+
+## Packages changes
+
+- `CronScheduler.Extensions` contains the basic generic code for Cron Scheduler.
+- `CronScheduler.AspNetCore` contains specific extensions for `AspNetCore` hosting.
+
+## Namespaces Changed
+
+`using CronScheduler.AspNetCore;` to `using CronScheduler.Extensions.Scheduler;`
+
