+++ README.md
@@ -9,10 +9,16 @@ It is much lighter than Quartz schedular and operates inside of any .NET Core Ge
 In addition `IStartupJob` was added to support async initialization before the IWebHost is ready to start. Sample project includes support for
 making sure that Database is created before the application starts.
 
-## .NET CLI
+## Install package for `AspNetCore` hosting .NET CLI
 
 ```bash
-    dotnet add package CronScheduler.AspNetCore --version 1.1.2
+    dotnet add package CronScheduler.AspNetCore
+```
+
+## Install package for `IHost` hosting .NET CLI
+
+```bash
+    dotnet add package CronScheduler.Extensions
 ```
 
 ## Uses Crontab format for Jobs/Tasks schedules
