+++ src/CronSchedulerApp/CronSchedulerApp.csproj
@@ -1,9 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <!--https://github.com/Microsoft/DockerTools/issues/34#issuecomment-481731741-->
-    <TargetFrameworks>netcoreapp2.2;netcoreapp3.0</TargetFrameworks>
-    <!--<TargetFramework>netcoreapp3.0</TargetFramework>-->
+    <TargetFramework>netcoreapp3.0</TargetFramework>
     <UserSecretsId>b790ca5d-c09d-4e9c-af99-c7b8c6b6210a</UserSecretsId>
     <AspNetCoreHostingModel>$(AspNetCoreHostingModel)</AspNetCoreHostingModel>
     <DockerDefaultTargetOS>$(DockerDefaultTargetOS)</DockerDefaultTargetOS>
@@ -12,26 +10,19 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.App" Condition="'$(TargetFramework)' == 'netcoreapp2.2'" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" />
     <PackageReference Include="Bet.AspNetCore" />
-    <PackageReference Include="Microsoft.Extensions.Http.Polly" />
-    <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" />
-  </ItemGroup>
-
-  <ItemGroup Condition="'$(TargetFramework)' == 'netcoreapp3.0'">
-
     <PackageReference Include="Microsoft.AspNetCore.Diagnostics.EntityFrameworkCore" />
     <PackageReference Include="Microsoft.AspNetCore.Identity.EntityFrameworkCore" />
     <PackageReference Include="Microsoft.AspNetCore.Identity.UI" />
-
     <PackageReference Include="Microsoft.EntityFrameworkCore.Relational" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" />
     <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" />
     <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" />
+    <PackageReference Include="Microsoft.Extensions.Http.Polly" />
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\CronScheduler\CronScheduler.csproj" />
+    <ProjectReference Include="..\CronScheduler.AspNetCore\CronScheduler.AspNetCore.csproj" />
   </ItemGroup>
 
 </Project>
