+++ src/CronSchedulerApp/Jobs/TorahQuoteJob.cs
@@ -2,7 +2,7 @@
 using System.Threading;
 using System.Threading.Tasks;
 
-using CronScheduler.AspNetCore;
+using CronScheduler.Extensions.Scheduler;
 
 using CronSchedulerApp.Services;
