+++ src/CronSchedulerWorker/CronSchedulerWorker.csproj
@@ -7,10 +7,10 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.Hosting" Version="3.0.0-preview5.19227.9" />
+    <PackageReference Include="Microsoft.Extensions.Hosting"/>
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\CronScheduler\CronScheduler.csproj" />
+    <ProjectReference Include="..\CronScheduler.Extensions\CronScheduler.Extensions.csproj" />
   </ItemGroup>
 </Project>
