+++ test/CronScheduler.UnitTest/TestJob.cs
@@ -2,7 +2,7 @@
 using System.Threading;
 using System.Threading.Tasks;
 
-using CronScheduler.AspNetCore;
+using CronScheduler.Extensions.Scheduler;
 
 using Microsoft.Extensions.Logging;
